Move source file related code into a separate lib
