+++ src/CMakeLists.txt
@@ -2,6 +2,7 @@ add_subdirectory(backtrace)
 add_subdirectory(base)
 add_subdirectory(ast)
 add_subdirectory(common)
+add_subdirectory(source_file)
 add_subdirectory(lexer)
 add_subdirectory(parser)
 add_subdirectory(analysis)
