+++ src/analysis/type_precheck.cpp
@@ -9,7 +9,7 @@
 #include "ast/decl.h"
 #include "ast/intrinsic.h"
 #include "ast/context.h"
-#include "lexer/token.h"
+#include "source_file/token.h"
 #include "compiler/compiler.h"
 #include <iostream>
 #include <set>
