+++ src/lexer/lexer.cpp
@@ -1,7 +1,7 @@
 #include "lexer/lexer.h"
 #include "base.h"
-#include "lexer/token.h"
-#include "lexer/source_file.h"
+#include "source_file/token.h"
+#include "source_file/source_file.h"
 #include <algorithm>
 #include <cctype>
 #include <regex>
