+++ src/lexer/token.cpp
@@ -1,4 +1,4 @@
-#include "lexer/token.h"
+#include "source_file/token.h"
 
 namespace tanlang {
