+++ src/source_file/CMakeLists.txt
@@ -1,0 +1,3 @@
+aux_source_directory(${PROJECT_SOURCE_DIR}/src/source_file SRCS)
+add_library(TanSourceFile STATIC ${SRCS})
+set_target_common_compile_options(TanSourceFile)
